Move and fix static_generator
This plugin will allow me to put arbitrary files in the same folder as
some content that I am producing and have it copied out into the output
folder. My changes prevent it from copying any file that is explicitly
mapped.

There is a bit of a philosophy question in the next logical feature,
excluding a file by filename. I'm not sure if I should include the .rst
or .md source files that go along with blog posts. I want to eventually
make all the stuff in git publically available so it may be another way
to acheive the goal of both source and content being available.